ekyc/live/requirements.txt
grpcio
opencv-python
numpy
dlib
protobuf
